refactor: icons export change
