+++ src/app/ar-assets/create/_components/StepContent/UploadQRCodeInsideImage.tsx
@@ -5,7 +5,7 @@ import { Center, Group, Stack } from '@/shared/components/common/Layout';
 import { Text } from '@/shared/components/common/Text';
 import { Title } from '@/shared/components/common/Title';
 import { SampleQrCodeImage } from '@/shared/components/features';
-import { IconUpload } from '@/shared/components/icons/IconUpload';
+import { IconUpload } from '@/shared/components/icons';
 
 export const UploadQRCodeInsideImage = () => {
   const [file, setFile] = useState<File | null>(null);
