+++ src/shared/components/common/Button/GoogleButton.tsx
@@ -1,4 +1,4 @@
-import { IconBrandGoogle } from '../../icons/IconBrandGoogle';
+import { IconBrandGoogle } from '../../icons';
 import { Button } from '.';
 
 type Props = {
