--- src/shared/components/icons/IconArrowBarToDown.tsx
@@ -1,3 +1,0 @@
-import { IconArrowBarToDown as TablerIconArrowBarToDown } from '@tabler/icons-react';
-
-export const IconArrowBarToDown = TablerIconArrowBarToDown;
