--- src/shared/components/icons/IconBrandGoogle.tsx
@@ -1,3 +1,0 @@
-import { IconBrandGoogle as TablerIcon } from '@tabler/icons-react';
-
-export const IconBrandGoogle = TablerIcon;
