--- src/shared/components/icons/IconCards.tsx
@@ -1,3 +1,0 @@
-import { IconCards as TablerCards } from '@tabler/icons-react';
-
-export const IconCards = TablerCards;
