--- src/shared/components/icons/IconCheck.tsx
@@ -1,3 +1,0 @@
-import { IconCheck as TablerIconCheck } from '@tabler/icons-react';
-
-export const IconCheck = TablerIconCheck;
