--- src/shared/components/icons/IconChevronLeft.tsx
@@ -1,3 +1,0 @@
-import { IconChevronLeft as TablerChevronLeft } from '@tabler/icons-react';
-
-export const IconChevronLeft = TablerChevronLeft;
