--- src/shared/components/icons/IconChevronRight.tsx
@@ -1,3 +1,0 @@
-import { IconChevronRight as TablerChevronRight } from '@tabler/icons-react';
-
-export const IconChevronRight = TablerChevronRight;
