--- src/shared/components/icons/IconExposurePlus1.tsx
@@ -1,3 +1,0 @@
-import { IconExposurePlus1 as TablerExposurePlus1 } from '@tabler/icons-react';
-
-export const IconExposurePlus1 = TablerExposurePlus1;
