--- src/shared/components/icons/IconLogout.tsx
@@ -1,3 +1,0 @@
-import { IconLogout as TablerIconLogout } from '@tabler/icons-react';
-
-export const IconLogout = TablerIconLogout;
