--- src/shared/components/icons/IconMicrophone.tsx
@@ -1,3 +1,0 @@
-import { IconMicrophone as TablerMicrophone } from '@tabler/icons-react';
-
-export const IconMicrophone = TablerMicrophone;
