--- src/shared/components/icons/IconPencil.tsx
@@ -1,3 +1,0 @@
-import { IconPencil as TablerIconPencil } from '@tabler/icons-react';
-
-export const IconPencil = TablerIconPencil;
