--- src/shared/components/icons/IconQrcode.tsx
@@ -1,3 +1,0 @@
-import { IconQrcode as TablerIconQrcode } from '@tabler/icons-react';
-
-export const IconQrcode = TablerIconQrcode;
