--- src/shared/components/icons/IconSpeakerphone.tsx
@@ -1,3 +1,0 @@
-import { IconSpeakerphone as TablerIconSpeakerphone } from '@tabler/icons-react';
-
-export const IconSpeakerphone = TablerIconSpeakerphone;
