--- src/shared/components/icons/IconUpload.tsx
@@ -1,3 +1,0 @@
-import { IconUpload as TablerIconUpload } from '@tabler/icons-react';
-
-export const IconUpload = TablerIconUpload;
