--- src/shared/components/icons/IconX.tsx
@@ -1,2 +1,0 @@
-import { IconX as TablerIcon } from '@tabler/icons-react';
-export const IconX = () => <TablerIcon />;
