+++ src/shared/components/icons/index.ts
@@ -1,12 +1,16 @@
-export { IconArrowBarToDown } from './IconArrowBarToDown';
-export { IconCards } from './IconCards';
-export { IconCheck } from './IconCheck';
-export { IconChevronLeft } from './IconChevronLeft';
-export { IconChevronRight } from './IconChevronRight';
-export { IconLogout } from './IconLogout';
-export { IconMicrophone } from './IconMicrophone';
-export { IconPencil } from './IconPencil';
-export { IconQrcode } from './IconQrcode';
-export { IconSpeakerphone } from './IconSpeakerphone';
-export { IconUpload } from './IconUpload';
-export { IconExposurePlus1 } from './IconExposurePlus1';
+export { IconArrowBarToDown } from '@tabler/icons-react';
+export { IconCards } from '@tabler/icons-react';
+export { IconCheck } from '@tabler/icons-react';
+export { IconChevronLeft } from '@tabler/icons-react';
+export { IconChevronRight } from '@tabler/icons-react';
+export { IconLogout } from '@tabler/icons-react';
+export { IconMicrophone } from '@tabler/icons-react';
+export { IconPencil } from '@tabler/icons-react';
+export { IconQrcode } from '@tabler/icons-react';
+export { IconSpeakerphone } from '@tabler/icons-react';
+export { IconUpload } from '@tabler/icons-react';
+export { IconExposurePlus1 } from '@tabler/icons-react';
+export { IconSun } from '@tabler/icons-react';
+export { IconMoon } from '@tabler/icons-react';
+export { IconBrandGoogle } from '@tabler/icons-react';
+export { IconX } from '@tabler/icons-react';
